Drop data for blacklisted metrics
As a companion to the hard max creation limit introduced in commit
8016b45, we propose adding a dynamic blacklist of metric prefixes.
Any data we receive for a metric whose name is prefixed with an entry in
the blacklist will be dropped. For example, if our blacklist contains
the prefix "foo", then new data for the metrics

    foo
    foo.bar
    foo.bar.spam

will be dropped, while we'll continue to accept data for the metrics:

    quux
    fo.bar
    etc

The blacklist can be controlled at runtime via an HTTP endpoint,
'_internal/blacklist'. It exposes a counter 'blacklistDropCount' of how
many metrics have been dropped because of the blacklist. The blacklist
prefixes are plain strings; no pattern or regex matching is done.

This functionality would be useful to mitigate the following outage
scenario: A user starts exploding the metric namespace, presumably by
accident. We've seen people do this by including any of the following in
their metric names:

- Dimensions of images (foo.bar.size640x480)
- Literal values of counters (foo.bar.counter.123)
- Hex checksums of some quantity (foo.bar.123456789ABCDEF)
- Any and all URL parameters they see
- Hostnames of Kubernetes pods that keep being restarted

We usually find out pretty quickly that something is going on, and the
explosion usually happens relatively deep inside the global metric
namespace. Having a hard limit on the creation of new metrics as in
8016b45 gives us time to react before the stores run out of disk space,
but sometimes it can be hard to track down the users behind the
production change that resulted in the explosion. They may, for example,
have gone to lunch after pushing their change out.

In those cases, it may be preferable to dig through some logs to find
the largest common prefix for the new metrics and simply block them
instead of running out of disk space or memory, and thus degrading both
the read and write services for everyone else.