Add an optional metric firehose
If the new config option 'firehose-udp-address' is specified, and is a
valid UDP address, we write

    "metric value timestamp"

to that address for every incoming data point. The writes have a
hardcoded 1 millisecond timeout. No error handling is provided for
dropped writes.

The use case for this is to have a sidecar tail that firehose and keep
an expiring count of writes to (metric, timestamp) pairs. That can
again be used to alert on repeated writes to the same metric for a given
timestamp, which we can use to poke users who may be corrupting their
data without even knowing it.

The following image is a helpful illustration of the firehose setup in
action:

    https://tinyurl.com/yctc4xnr